[AWN-39270] Add http hostname to flow json log output
This entry is being added to remove the need for YAF and the subsequent
parsing of its ipfix-format output, which has become a performance
bottleneck.

Testing:
Basic manual testing was done with some simple HTTP flows, non-HTTP TCP
flows and non-TCP flows. AddressSanitizer was used to test for memory
leaks.