Add active-flow timeout to replicate yaf behaviour
Introduced configurable "active" and "emergency-active" flow timeouts
for TCP, UDP, ICMP and Default (i.e. everything else) protocols which
make suricata behave more like yaf: When active timeouts are
configured, the state of flows that exist for longer than the
configured timeout are dropped, and any subsequent packets for that
flow are seen as a new flow. This behaviour is disabled by default.
The existing behaviour can also be explicitly configured by setting the
active and emergency-active timeouts to zero.

If flow logging is enabled, the flow state is logged upon expiration of
the active-timeout.

The following testing was done with and without flow logging enabled:
- Verified default behaviour is as before (i.e. no active timeout
    config in /etc/suricata/suricata.yaml).
- Configured different active timeouts for each protocol.
- Explicity disabled active timeouts by setting them to zero.