[AWN-63253] Ensure tcp_init_flags are initialized
The tcp_init_flags_* params are part of a TcpSession
object which is allocated out of a pool which is
initially cleared to 0.
However, if a TcpSession is released, and then
reallocated, then it's assumed to be lazily
initialized.
In this case, however, the flags corresponding
to the opposite direction of the SYN packet were
not cleared to 0, which meant they contained the
previous/stale flags from the last use.