Changed ntohs to __constant_htons to keep the eBPF validator happy.
Also added a "--debug" flag to build-xdp.sh which will generate a BPF with debug symbols. llvm-objdump can then be used to generate an annotated assembly listing, which can be used in conjunction with syslog messages to determine what happened when the kernel refuses to load a given BPF program.