[SSC-826] Fix AFPv3 socket ref count
When using AFPacket V3, we we can use
the packet data right from the ring (i.e.,
zero-copy) we increment the socket ref
to ensure the NIC is brought down while
there are still references to the rx-queue.
However, we don't ever decrement those
references.
This add a ref decrement upon freeing
the packet.